backend/requirements.txt
supabase
pydantic-settings
pydantic
pydantic-core
pydantic-settings
google-genai